extend string hash

+++ backend/src/parallel_RA_inc.h
@@ -56,7 +56,7 @@ inline u64 number_to_datum(i64 number) {
 const auto n2d = number_to_datum;
 
 inline u64 string_to_datum(std::string str) {
-    u32 str_hash = string_hash(str);
+    u64 str_hash = string_hash(str);
     return (str_hash & ~tag_mask) | (str_tag << tag_position);
 }
 const auto s2d = string_to_datum;
